deleteDebt implemented with prior confirmation to accept

+++ src/components/ConfirmDelete.tsx
@@ -1,15 +1,16 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 "use client";
 import { useState } from "react";
-import { postUser, patchUser, deleteUser } from "../lib/api";
-import type { NewUser, User } from "../types";
+import { deleteUser, deleteDebt } from "../lib/api";
+import type { Debt, User } from "../types";
 
 interface ConfirmDeleteProps {
-    userDelete?: User;
-    setUserDelete?: (user?: User) => void;
     setVisibleConfirmDelete?: (visible: boolean) => void;
+    userDelete?: User;
+    debtDelete?: Debt;
 }
 
-export default function ConfirmDelete({ userDelete, setUserDelete, setVisibleConfirmDelete }: ConfirmDeleteProps) {
+export default function ConfirmDelete({ setVisibleConfirmDelete, userDelete, debtDelete }: ConfirmDeleteProps) {
     
     
     const [loading, setLoading] = useState(false);
@@ -21,17 +22,19 @@ export default function ConfirmDelete({ userDelete, setUserDelete, setVisibleCon
         setMessage(null);
 
         try {
-            if(userDelete) {
+            if(userDelete && !debtDelete) {
                 await deleteUser(userDelete._id);
+            } 
+            if(!userDelete && debtDelete) {
+                await deleteDebt(debtDelete._id);
             } 
             // console.log("Created user:", response);
 
-            setMessage("User deleted");
-            setUserDelete?.(undefined);
+            setMessage("Deleted successfully");
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         } catch(err: any) {
-            console.error("Error creating user:", err);
-            setMessage(`X ${err.message || "Failed to create user"}`);
+            console.error("Error deleting user:", err);
+            setMessage(`X ${err.message || "Failed to delete"}`);
         } finally {
             setVisibleConfirmDelete?.(false);
             setLoading(false);
@@ -41,21 +44,53 @@ export default function ConfirmDelete({ userDelete, setUserDelete, setVisibleCon
     return (
         <div className="flex flex-col p-2 max-w-md">
             <h2 className="text-3xl font-bold text-gray-800 dark:text-gray-100 mb-4">
-                Delete User
+                Delete 
+                {userDelete && " User"} 
+                {debtDelete && " Debt"}
             </h2>
             
-            <form onSubmit={handleSubmit} className="flex flex-col gap-1.5 border rounded-lg p-2 bg-blue-300 text-gray-800">
-                <label className="input-label">Email: {userDelete?.email}</label>
-                <label className="input-label">Phone: {userDelete?.phone}</label>
-                <label className="input-label">Name: {userDelete?.name}</label>
-                <label className="input-label">Enabled: {userDelete?.enabled}</label>
-                <label className="input-label">Deleted: {userDelete?.deleted}</label>
+            <form onSubmit={handleSubmit} className={`flex flex-col gap-1.5 border rounded-lg p-2 ${userDelete && "bg-blue-300"} ${debtDelete && "bg-green-300"} text-gray-800`}>
+                {(userDelete?._id && !debtDelete?._id) && ( 
+                    <>
+                        <label className="input-label">Email: {userDelete?.email}</label>
+                        <label className="input-label">Phone: {userDelete?.phone}</label>
+                        <label className="input-label">Name: {userDelete?.name}</label>
+                        <label className="input-label">Enabled: {userDelete?.enabled}</label>
+                        <label className="input-label">Deleted: {userDelete?.deleted}</label>
+                    </>
+                ) }
+                {(!userDelete?._id && debtDelete?._id) && ( 
+                    <>
+                        {/* <label className="input-label">ID Creditor: {debtDelete?.id_user_creditor}</label> */}
+                        <label className="input-label">
+                            ID Creditor: {typeof debtDelete?.id_user_creditor === "object"
+                            ? (debtDelete.id_user_creditor as any)._id
+                            : debtDelete?.id_user_creditor}
+                        </label>
+                        {/* <label className="input-label">ID Debtor: {debtDelete?.id_user_debtor}</label> */}
+                        <label className="input-label">
+                            ID Debtor: {typeof debtDelete?.id_user_debtor === "object"
+                            ? (debtDelete.id_user_debtor as any)._id
+                            : debtDelete?.id_user_debtor}
+                        </label>
+                        <label className="input-label">Detail: {debtDelete?.detail}</label>
+                        <label className="input-label">Amount: {debtDelete?.amount}</label>
+                        <label className="input-label">Dolar Google: {debtDelete?.dolar_google}</label>
+                        <label className="input-label">Status: {debtDelete?.status}</label>
+                        <label className="input-label">Date Debt: {debtDelete?.date_debt}</label>
+                        <label className="input-label">Date Due: {debtDelete?.date_due}</label>
+                        <label className="input-label">Currency: {debtDelete?.currency}</label>
+                        <label className="input-label">Enabled: {debtDelete?.enabled}</label>
+                        <label className="input-label">Deleted: {debtDelete?.deleted}</label>
+                    </>
+                ) }
+
                 <button
                     type="submit"
                     disabled={loading}
-                    className="bg-gray-500 text-white py-2 m-auto w-30 rounded-lg hover:bg-gray-700 transition cursor-pointer"
+                    className="bg-red-500 text-white py-2 m-auto w-30 rounded-lg hover:bg-gray-700 transition cursor-pointer"
                 >
-                    Confirm Delete User   
+                    Confirm Delete   
                 </button>
             </form>
